inference_pipeline/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-image>=0.21.0
pyyaml>=6.0
tqdm>=4.65.0
streamlit>=1.28.0
plotly>=5.15.0
streamlit-image-comparison>=0.0.4